Add option to signal parent process when ptys are ready
Adds a --signal-parent (-s) command line option which, when specifies,
causes nulltty to send the designated signal to its parent process when
the pseudoterminals are ready to be opened. The intended purpose for
this is in nulltty's "make check" test suite, so that the controlling
test program knows when it is safe to begin.