Preliminary support for building on OpenBSD
Supports OpenBSD by allowing the use of openpty() in place of
posix_openpt(), and select() in place of pselect().

But we aren't done with the port yet: there is an autoconf script
glitch preventing consistent addition of -lutil on OpenBSD, and the
"make check" test program hangs after it runs.