Reduce test message size as a temporary fix for OpenBSD blocking
Our test process doesn't take into account that the write buffers for
the ptys may overflow if we write an amount of data greater than the
buffer size, causing the test program to hang indefinitely in the
current design.

As a temporary workaround we can reduce the amount of data sent, where
1024 is too much on OpenBSD 5.2 amd64, but a better solution would be
to write data in a select loop.